+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 - To improve performance, FileTypeMetadata.Exists() should use base Python rather than DataManagement.Files.File.Exists() ([#31](https://github.com/jjrob/MGET/issues/31))
 - RuntimeError: Failed to open a variable named "sea_ice_fraction" in netCDF file ... from NASA Earthdata GHRSST GAMSSA_28km-ABOM-L4-GLOB-v01 granules. Detailed error information: KeyError: 'sea_ice_fraction'. ([#32](https://github.com/jjrob/MGET/issues/32))
 - FastMarchingDistanceGrid fails with NameError: name 'GridSlice' is not defined ([#34](https://github.com/jjrob/MGET/issues/34))
+- When copernicusmarine 2.1.0 is installed, MGET functions fail with ModuleNotFoundError: No module named 'copernicusmarine.download_functions.download_arco_series' ([#36](https://github.com/jjrob/MGET/issues/36))
 
 ## [v3.2.0](https://github.com/jjrob/MGET/releases/tag/v3.2.0) - 2025-02-16